Don't retain any clear screen sequence variants.
(Simpler, replaces previous patch)

Elide all clear screen escape sequences from the scrollback buffer,
partial or complete. zshell, for example, emits ^[[J on startup, which
causes the scrolled down portion of the screen to be cleared when
scrolling to the beginning of the buffer and less than the entire screen
is scrolled.